==> unsupervised_notebooks/04_analyze_train_results.ipynb <==
# Analyze Train Results
## Dependencies

# !pip install -r requirements.txt

import pandas as pd
import numpy as np
import joblib
import os
import json

filepath = '/home/datallah/datallah-jaymefis-gibsonce/'
sample_size = 'one'
random_state = 42

d_path = filepath + 'documentation/'

## Import Results

# load df
df = pd.read_csv(d_path + 'train_scores.csv', sep = '|')

# flatten score json's into df
score_keys = json.loads(df.scores[0]).keys()
for key in score_keys:
    df[key] = df.scores.apply(lambda x: json.loads(x)[key])
df.head()

# sort results by accuracy
df['mean_test_accuracy2'] = 1 - df.mean_test_accuracy
df['mean_test_accuracy3'] = df[['mean_test_accuracy', 'mean_test_accuracy2']].max(axis = 1)
df[['model_nm', 'matrix_nm', 'params', 'mean_test_accuracy3']].sort_values(
    by = 'mean_test_accuracy3', ascending = False).head(10)

df[['model_nm', 'matrix_nm', 'params', 'mean_test_accuracy3', 'mean_test_adjusted_rand_score']].sort_values(
    by = ['mean_test_adjusted_rand_score', 'mean_test_accuracy3'], ascending = False).head(10)

print(df.sort_values(by = ['mean_test_accuracy3'], ascending = False).params.iloc[0])

df[['model_nm', 'matrix_nm', 'mean_test_accuracy3', 'mean_test_adjusted_rand_score']].groupby(
    by = ['model_nm', 'matrix_nm']).mean().sort_values(
    by = ['mean_test_adjusted_rand_score', 'mean_test_accuracy3'], ascending = False)
