==> unsupervised_notebooks/05_cluster_tune_BGMM_lower_dim.ipynb <==
# BGMM
## Dependencies

# !pip install -r requirements.txt

import pandas as pd
import numpy as np
import joblib
import os
import json
from collections import Counter
import re
import tqdm
from datetime import datetime
import multiprocessing

### sklearn dependencies
import sklearn
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.preprocessing import normalize
from sklearn.decomposition import TruncatedSVD
from sklearn.mixture import BayesianGaussianMixture
from sklearn.model_selection import GridSearchCV
from scipy import sparse

### text preprocessing dependencies
import nltk
from nltk.tokenize.casual import TweetTokenizer
nltk.download('wordnet')
from nltk.stem.wordnet import WordNetLemmatizer

### gensim dependencies
from gensim.models import Word2Vec, KeyedVectors
from gensim.test.utils import datapath

filepath = '/home/datallah/datallah-jaymefis-gibsonce/'
random_state = 42
stop = {'a', 'about', 'above', 'after', 'again', 'against', 'ain',
        'all', 'am', 'an', 'and', 'any', 'are', 'aren', "aren't",
        'as', 'at', 'be', 'because', 'been', 'before', 'being',
        'below', 'between', 'both', 'but', 'by', 'can', 'couldn',
        "couldn't", 'd', 'did', 'didn', "didn't", 'do', 'does',
        'doesn', "doesn't", 'doing', 'don', "don't", 'down', 'during',
        'each', 'few', 'for', 'from', 'further', 'had', 'hadn',
        "hadn't", 'has', 'hasn', "hasn't", 'have', 'haven', "haven't",
        'having', 'here', 'how', 'i', 'if', 'in', 'into', 'is', 'isn',
        "isn't", 'it', "it's", 'its', 'itself', 'just', 'll', 'm',
        'ma', 'me', 'mightn', "mightn't", 'more', 'most', 'mustn',
        "mustn't", 'my', 'myself', 'needn', "needn't", 'no', 'nor',
        'not', 'now', 'o', 'of', 'off', 'on', 'once', 'only', 'or',
        'other', 'our', 'ours', 'ourselves', 'out', 'over', 'own',
        're', 's', 'same', 'shan', "shan't", 'should', "should've",
        'shouldn', "shouldn't", 'so', 'some', 'such', 't', 'than',
        'that', "that'll", 'the', 'their', 'theirs', 'them', 'themselves',
        'then', 'there', 'these', 'they', 'this', 'those', 'through',
        'to', 'too', 'under', 'until', 'up', 've', 'very', 'was', 'wasn',
        "wasn't", 'we', 'were', 'weren', "weren't", 'what', 'when', 'where',
        'which', 'while', 'who', 'whom', 'why', 'will', 'with', 'won',
        "won't", 'wouldn', "wouldn't", 'y', 'you', "you'd", "you'll",
        "you're", "you've", 'your', 'yours', 'yourself', 'yourselves'}

size = 'one'
train = pd.read_csv(filepath + f'samples/train_{size}_million.csv').rename(
    columns = {' response_text': 'response_text', ' op_gender': 'op_gender'}).dropna()
val   = pd.read_csv(filepath + f'samples/validate_{size}_million.csv').rename(
    columns = {' response_text': 'response_text', ' op_gender': 'op_gender'}).dropna()
test  = pd.read_csv(filepath + f'samples/test_{size}_million.csv').rename(
    columns = {' response_text': 'response_text', ' op_gender': 'op_gender'}).dropna()

X_train = train[train.source == 'TED'].response_text
y_train = train[train.source == 'TED'].op_gender
X_val = val[val.source == 'TED'].response_text
y_val = val[val.source == 'TED'].op_gender
X_test  = test[test.source == 'TED'].response_text
y_test  = test[test.source == 'TED'].op_gender

del train, val, test

X_train.shape

## Load TF-IDF

# tokenize functions
# remove links as these will only apply to specific responses
def rm_link(text):
    return re.sub(r'https?://\S+', '', text)

# remove punctuation
def rm_punct(text):
    return re.sub(r'[^\w\s]', '', text)

# create class that lemmatizes tweet tokens
# this will be used when creating the term matrix
class LemmaTokenizer(object):
    def __init__(self):
        self.wnl = WordNetLemmatizer()
        self.tt = TweetTokenizer(preserve_case=False, reduce_len=True,
                                 strip_handles=True, match_phone_numbers=False)
    def __call__(self, docs):
        return [self.wnl.lemmatize(t) for t in self.tt.tokenize(rm_link(rm_punct(docs))) if t not in stop]
# creates a term matrix
def train_vectorizer(text_data, vectorizer=CountVectorizer, tokenizer=LemmaTokenizer(),
                     ngram_range_lower = 1, ngram_range_upper = 1, min_df = 1):
    """
    Trains a vectorizer on the provided text data and returns the vectorizer instance,
    the document-term matrix, and the feature names.

    Parameters:
    - text_data: List of text documents to be vectorized.
    - vectorizer: Vectorizer class to be used for text vectorization. Defaults to CountVectorizer.
    - tokenizer: Tokenizer class to be used for tokenizing the text documents. Defaults to TweetTokenizer.
    - ngram_range_lower: What's the minimum length of n-grams we want.
    - ngram_range_upper: What's the maximum length of n-grams we want.
    - min_df: Minimum data frequency.
    - max_df: Maximum data frequency.

    Returns:
    - instance: The trained vectorizer instance.
    - matrix: The document-term matrix resulting from fitting the vectorizer on `text_data`.
    - features: An array of feature names generated by the vectorizer.
    """
    # Initialize the vectorizer with specified configurations
    instance = vectorizer(
        strip_accents=None,  # Do not strip accents
        lowercase=True,  # Do not convert characters to lowercase
        tokenizer=tokenizer,  # Use the tokenize method of the tokenizer instance
        token_pattern=None,  # Since a tokenizer is provided, token_pattern is not used
        ngram_range=(ngram_range_lower, ngram_range_upper),  # Consider only single words (1-grams)
        min_df=min_df,  # Minimum document frequency for filtering terms
        max_features=None  # No limit on the number of features
    )

    # Fit the vectorizer on the provided text data and transform the data into a matrix
    vector = instance.fit(text_data)

    return vector, instance

tfidf_v, instance = train_vectorizer(text_data = X_train,
                           vectorizer = TfidfVectorizer,
                           ngram_range_lower = 1,
                           ngram_range_upper = 3,
                           min_df = 5)

features = '|\n|'.join(instance.get_feature_names())
file = open('/home/datallah/datallah-jaymefis-gibsonce/bgmm/features.txt', "w")
file.write(features)
file.close()  # to change file access modes

# normalize and truncate

print("Transforming X_train")
tfidf_m = tfidf_v.transform(X_train)
sparse.save_npz('/home/datallah/datallah-jaymefis-gibsonce/bgmm/tfidf_m.npz', tfidf_m)
del X_train
# truncate
svd = TruncatedSVD(n_components = 25, random_state = random_state)
svd.fit(normalize(tfidf_m))
tfidf_m = svd.transform(normalize(tfidf_m))
np.save('/home/datallah/datallah-jaymefis-gibsonce/bgmm/tfidf_trunc_low_dim.npy', tfidf_m)

print("Transforming X_val")
tfidf_m_val = tfidf_v.transform(X_val)
del X_val
sparse.save_npz('/home/datallah/datallah-jaymefis-gibsonce/bgmm/tfidf_m_val.npz', tfidf_m_val)
tfidf_m_val = svd.transform(normalize(tfidf_m_val))
np.save('/home/datallah/datallah-jaymefis-gibsonce/bgmm/tfidf_trunc_val_low_dim.npy', tfidf_m_val)

print("Transforming X_test")
tfidf_m_test = tfidf_v.transform(X_test)
del X_test
sparse.save_npz('/home/datallah/datallah-jaymefis-gibsonce/bgmm/tfidf_m_test.npz', tfidf_m_test)
tfidf_m_test = svd.transform(normalize(tfidf_m_test))
np.save('/home/datallah/datallah-jaymefis-gibsonce/bgmm/tfidf_trunc_test_low_dim.npy', tfidf_m_test)

tfidf_m.shape

## Train Base Model

# init variables
best_model = None
best_log_likelihood = -np.inf
log_lik_lst = []
time_start_lst = []
time_end_lst = []
max_components = np.arange(1, 30 + 1)
breaker = 0

# iterate over different numbers of components
for n_components in tqdm.tqdm(max_components):
    time_start_lst.append(datetime.now())
    # Initialize and fit the BGMM with current number of components
    bgmm = BayesianGaussianMixture(n_components = n_components,
                                   random_state = random_state, 
                                   max_iter = 1000,
                                   warm_start = True)
    bgmm.fit(tfidf_m)
    
    # Calculate log likelihood of the current model
    log_likelihood = bgmm.score(tfidf_m_val)
    log_lik_lst.append(log_likelihood)
    print(f"Number of components: {n_components}, Log likelihood: {log_likelihood}")
    
    time_end_lst.append(datetime.now())
    # Check if log likelihood is better than the best so far
    if log_likelihood > best_log_likelihood:
        best_log_likelihood = log_likelihood
        best_model = bgmm
        breaker = 0
    elif breaker < 3:
        breaker ++ 1
        # If log likelihood starts to dip, stop iterating
    else:
        break

max_components = np.arange(1, 30 + 1)
log_lik_lst = [196.57135188307578, 243.11340103156016, 260.9960548937567, 290.687483564508, 302.5393462420144, 311.82450073233025, 319.2109229524175, 316.05980845011544, 324.16009071553856, 329.4315077583328, 329.7545834976157, 335.2015776342216, 338.5252807338984, 340.07432639411326, 344.4102559634466, 346.4648811376254, 349.14685280030614, 349.6416119071051, 351.64877838142047, 351.8723286693082, 353.37432998827825]
times_str = ['00:00:00', '00:00:02', '00:01:36', '00:05:31', '00:08:36', '00:14:30', '00:18:57', '00:25:00', '00:28:35', '00:36:06', '00:43:32', '00:49:25', '01:01:46', '01:20:33', '01:48:10', '02:07:35', '02:32:11', '03:14:35', '03:42:39', '04:29:00', '04:57:43', '05:23:08']

times = [datetime.strptime(time_str, '%H:%M:%S') for time_str in times_str]
time_diff_lst = [(times[i + 1] - times[i]).total_seconds() for i in range(len(times) - 1)]
print(time_diff_lst)

i = len(log_lik_lst) - 1
# create and write sensitivity df
sens_df = pd.DataFrame()
sens_df['n_componenets'] = max_components[0:i]
sens_df['max_log_likelihood'] = log_lik_lst[0:i]
sens_df['train_time'] = time_diff_lst[0:i]
sens_df
# write df
sens_df.to_csv('/home/datallah/datallah-jaymefis-gibsonce/bgmm/sensitivity.csv', index = False) 

## Train Base Model (Larger step)

!python bgmm_multi_train_low_dim.py
