==> unsupervised_notebooks/06_final_BGM.ipynb <==
# Explore BGMM

## Dependencies

import pandas as pd
import numpy as np
import joblib
import os
import json
from collections import Counter
import re
import tqdm
from datetime import datetime

### sklearn dependencies
import sklearn
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.preprocessing import normalize
from sklearn.decomposition import TruncatedSVD
from sklearn.mixture import BayesianGaussianMixture
from sklearn.model_selection import GridSearchCV
from scipy import sparse

### text preprocessing dependencies
import nltk
from nltk.tokenize.casual import TweetTokenizer
nltk.download('wordnet')
from nltk.stem.wordnet import WordNetLemmatizer

### gensim dependencies
from gensim.models import Word2Vec, KeyedVectors
from gensim.test.utils import datapath

filepath = '/home/datallah/datallah-jaymefis-gibsonce/'
random_state = 42

size = 'one'
train = pd.read_csv(filepath + f'samples/train_{size}_million.csv').rename(
    columns = {' response_text': 'response_text', ' op_gender': 'op_gender'}).dropna()
val   = pd.read_csv(filepath + f'samples/validate_{size}_million.csv').rename(
    columns = {' response_text': 'response_text', ' op_gender': 'op_gender'}).dropna()
test  = pd.read_csv(filepath + f'samples/test_{size}_million.csv').rename(
    columns = {' response_text': 'response_text', ' op_gender': 'op_gender'}).dropna()

X_train = train[train.source == 'TED'].response_text
y_train = train[train.source == 'TED'].op_gender
X_val = val[val.source == 'TED'].response_text
y_val = val[val.source == 'TED'].op_gender
X_test  = test[test.source == 'TED'].response_text
y_test  = test[test.source == 'TED'].op_gender

del train, val, test

tfidf_m = np.load('/home/datallah/datallah-jaymefis-gibsonce/bgmm/tfidf_trunc.npy')
tfidf_m_val = np.load('/home/datallah/datallah-jaymefis-gibsonce/bgmm/tfidf_trunc_val.npy')
tfidf_m_test = np.load('/home/datallah/datallah-jaymefis-gibsonce/bgmm/tfidf_trunc_test.npy')

sens_df = pd.read_csv('/home/datallah/datallah-jaymefis-gibsonce/bgmm/sensitivity.csv')
sens_df

## Fit Model with Best Base Params

n_components = 20

bgmm = BayesianGaussianMixture(n_components = n_components,
                               random_state = random_state, 
                               max_iter = 1000)
bgmm.fit(tfidf_m)

## Analyze Component Gender Makeup

train_preds = bgmm.predict(tfidf_m)
train_preds_df = pd.DataFrame({'preds' : train_preds, 
                               'label' : np.where(y_train == 'W', 0, 1)})
train_preds_df.sample(5)

components = np.sort(train_preds_df.preds.unique())
men_ratio = train_preds_df.groupby('preds').apply(lambda x: sum(x.label)/len(x))
cnts = train_preds_df.groupby('preds')['label'].count()
grpd = pd.DataFrame({'component'   : components,
                     'men_ratio'   : men_ratio,
                     'women_ratio' : 1 - men_ratio,
                     'comp_cnts'   : cnts})
grpd.sort_values(by = 'men_ratio', ascending = False)

sig_comp_df = grpd[(grpd.men_ratio >= 0.6) | (grpd.women_ratio >= 0.6)].sort_values(by = 'men_ratio', ascending = False)
sig_comps = list(sig_comp_df.component.unique())
print(sig_comps)

posterior_probs = bgmm.predict_proba(tfidf_m)

sig_docs = {}
n_top_docs = sig_comp_df.comp_cnts.min()
for comp in sig_comps:
    # Get the indices of the top n documents for the current component
    top_docs_idx = list(np.argsort(posterior_probs[:, comp])[::-1][:n_top_docs])
    # Retrieve the top n documents for the current component
    top_docs = [X_train.iloc[idx] for idx in top_docs_idx]
    sig_docs[comp] = top_docs

# example doc
for key in sig_comps:
    print(f'Examples for component {str(key)}:')
    for i in range(10):
        print(f'{str(i + 1)}: {sig_docs[key][i]}')
    print('\n')

## Find Significant Words

# import tf-idf
full_train_tfidf = normalize(sparse.load_npz('/home/datallah/datallah-jaymefis-gibsonce/bgmm/tfidf_m.npz'))

file = open('/home/datallah/datallah-jaymefis-gibsonce/bgmm/features.txt', "r")
features = file.read().split('|\n|')
file.close()
features[:5]

sig_vec = {}
n_top_docs = sig_comp_df.comp_cnts.min()
for comp in sig_comps:
    # Get the indices of the top n documents for the current component
    top_docs_idx = list(np.argsort(posterior_probs[:, comp])[::-1][:n_top_docs])
    # Retrieve the top n documents for the current component
    top_docs = [full_train_tfidf[idx] for idx in top_docs_idx]
    sig_vec[comp] = top_docs

sig_terms = {}
# example doc
for key in sig_comps:
    print(f'Significant terms for component {str(key)}:')
    # find significant terms by average weight
    avg_tfidf_scores = sparse.csr_matrix.mean(np.array(sig_vec[key]), axis = 0)
    top_term_indx = np.argsort(avg_tfidf_scores.toarray()[0])[::-1]
    n_top_terms = 50
    top_terms = [features[idx] for idx in top_term_indx[:n_top_terms]]
    sig_terms[key] = top_terms
    
    print(top_terms)

# remove redundant values to exclude special values
top_terms = [term for lst in sig_terms.values() for term in lst]
term_counts = Counter(top_terms)
exclude_terms = {term for term, count in term_counts.items() if count > 1}
for key in sig_comps:
    print(f'Significant (non-redundant) terms for component {str(key)}:')
    temp_lst = [term for term in sig_terms[key] if term not in exclude_terms]
    print(temp_lst)

## Analyze Component Gender Makeup (Test)

test_preds = bgmm.predict(tfidf_m_test)
test_preds_df = pd.DataFrame({'preds' : test_preds, 
                              'label' : np.where(y_test == 'W', 0, 1)})
test_preds_df.sample(5)

components = np.sort(test_preds_df.preds.unique())
men_ratio = test_preds_df.groupby('preds').apply(lambda x: sum(x.label)/len(x))
cnts = test_preds_df.groupby('preds')['label'].count()
grpd = pd.DataFrame({'component'   : components,
                     'men_ratio'   : men_ratio,
                     'women_ratio' : 1 - men_ratio,
                     'comp_cnts'   : cnts})
grpd.sort_values(by = 'men_ratio', ascending = False)

sig_comp_df = grpd[(grpd.men_ratio >= 0.6) | (grpd.women_ratio >= 0.6)].sort_values(by = 'men_ratio', ascending = False)
sig_comps = list(sig_comp_df.component.unique())
print(sig_comps)

posterior_probs = bgmm.predict_proba(tfidf_m_test)

sig_docs = {}
n_top_docs = sig_comp_df.comp_cnts.min()
for comp in sig_comps:
    # Get the indices of the top n documents for the current component
    top_docs_idx = list(np.argsort(posterior_probs[:, comp])[::-1][:n_top_docs])
    # Retrieve the top n documents for the current component
    top_docs = [X_test.iloc[idx] for idx in top_docs_idx]
    sig_docs[comp] = top_docs

# example doc
for key in sig_comps:
    print(f'Examples for component {str(key)}:')
    for i in range(10):
        print(f'{str(i + 1)}: {sig_docs[key][i]}')
    print('\n')

## Find Significant Words

# import tf-idf
full_test_tfidf = normalize(sparse.load_npz('/home/datallah/datallah-jaymefis-gibsonce/bgmm/tfidf_m_test.npz'))

# file = open('/home/datallah/datallah-jaymefis-gibsonce/bgmm/features_test.txt', "r")
# features = file.read().split('|\n|')
# file.close()
# features[:5]

sig_vec = {}
n_top_docs = sig_comp_df.comp_cnts.min()
for comp in sig_comps:
    # Get the indices of the top n documents for the current component
    top_docs_idx = list(np.argsort(posterior_probs[:, comp])[::-1][:n_top_docs])
    # Retrieve the top n documents for the current component
    top_docs = [full_test_tfidf[idx] for idx in top_docs_idx]
    sig_vec[comp] = top_docs

sig_terms = {}
# example doc
for key in sig_comps:
    print(f'Significant terms for component {str(key)}:')
    # find significant terms by average weight
    avg_tfidf_scores = sparse.csr_matrix.mean(np.array(sig_vec[key]), axis = 0)
    top_term_indx = np.argsort(avg_tfidf_scores.toarray()[0])[::-1]
    n_top_terms = 50
    top_terms = [features[idx] for idx in top_term_indx[:n_top_terms]]
    sig_terms[key] = top_terms
    
    print(top_terms)

# remove redundant values to exclude special values
top_terms = [term for lst in sig_terms.values() for term in lst]
term_counts = Counter(top_terms)
exclude_terms = {term for term, count in term_counts.items() if count > 1}
for key in sig_comps:
    print(f'Significant (non-redundant) terms for component {str(key)}:')
    temp_lst = [term for term in sig_terms[key] if term not in exclude_terms]
    print(temp_lst)
